using theme via Emotion for type styles - various TS-driven refactoring
