+++ plugins/plugin-chart-word-cloud/package.json
@@ -33,6 +33,7 @@
     "@types/react": "^16.3.0",
     "d3-cloud": "^1.2.5",
     "d3-scale": "^3.0.1",
+    "emotion-theming": "^10.0.27",
     "encodable": "^0.3.3"
   },
   "peerDependencies": {
