+++ plugins/plugin-chart-word-cloud/src/plugin/index.ts
@@ -2,7 +2,7 @@ import { t } from '@superset-ui/translation';
 import { ChartMetadata, ChartPlugin } from '@superset-ui/chart';
 import buildQuery from './buildQuery';
 import { WordCloudFormData } from '../types';
-import transformProps from '../legacyPlugin/transformProps';
+import transformProps from '@superset-ui/plugin-chart-word-cloud/src/legacyPlugin/transformProps';
 import thumbnail from '../images/thumbnail.png';
 import controlPanel from './controlPanel';
