+++ plugins/plugin-chart-word-cloud/src/plugin/transformProps.tsx
